+++ gulpfile.js
@@ -1,20 +1,30 @@
 "use strict";
 
 var gulp = require("gulp");
+
+var sass = require("gulp-sass");
 var plumber = require("gulp-plumber");
 var sourcemap = require("gulp-sourcemaps");
+
 var rename = require("gulp-rename");
-var sass = require("gulp-sass");
-var postcss = require("gulp-postcss");
-var autoprefixer = require("autoprefixer");
-var csso = require("gulp-csso");
+var del = require("del");
 var imagemin = require("gulp-imagemin");
 var webp = require("gulp-webp");
 var svgstore = require("gulp-svgstore");
+
+var postcss = require("gulp-postcss");
+var autoprefixer = require("autoprefixer");
+var csso = require("gulp-csso");
+
 var posthtml = require("gulp-posthtml");
 var include = require("posthtml-include");
-var del = require("del");
+var htmlmin = require("gulp-htmlmin");
+
+var concat = require("gulp-concat");
+var uglify = require("gulp-uglify");
+
 var server = require("browser-sync").create();
+
 
 gulp.task("css", function () {
   return gulp.src("source/sass/style.scss")
@@ -29,6 +39,23 @@ gulp.task("css", function () {
     .pipe(sourcemap.write("."))
     .pipe(gulp.dest("build/css"))
     .pipe(server.stream());
+});
+
+gulp.task("html", function () {
+  return gulp.src("source/*.html")
+    .pipe(posthtml([
+      include()
+    ]))
+    .pipe(htmlmin({ collapseWhitespace: true }))
+    .pipe(gulp.dest("build"))
+});
+
+gulp.task("scripts", function() {
+  return gulp.src("source/js/*.js")
+    .pipe(concat("main.js"))
+    .pipe(uglify())
+    .pipe(rename("main.min.js"))
+    .pipe(gulp.dest("build/js"));
 });
 
 gulp.task("images", function () {
@@ -56,14 +83,6 @@ gulp.task("sprite", function () {
     .pipe(gulp.dest("build/img"))
 });
 
-gulp.task("html", function () {
-  return gulp.src("source/*.html")
-    .pipe(posthtml([
-      include()
-    ]))
-    .pipe(gulp.dest("build"))
-});
-
 gulp.task("copy", function () {
   return gulp.src([
     "source/fonts/**/*.{woff,woff2}",
@@ -84,6 +103,7 @@ gulp.task("build", gulp.series(
   "clean",
   "copy",
   "css",
+  "scripts",
   "sprite",
   "html",
 ));
@@ -100,6 +120,7 @@ gulp.task("server", function () {
   gulp.watch("source/sass/**/*.scss", gulp.series("css"));
   gulp.watch("source/img/*.svg", gulp.series("sprite", "html", "refresh"));
   gulp.watch("source/*.html", gulp.series("html", "refresh"));
+  gulp.watch("source/js/*.js", gulp.series("scripts", "refresh"));
 });
 
 gulp.task("refresh", function (done) {
