+++ package.json
@@ -8,6 +8,7 @@
     "autoprefixer": "^9.7.6",
     "browser-sync": "2.26.x",
     "gulp": "4.0.x",
+    "gulp-concat": "^2.6.1",
     "gulp-csso": "^4.0.1",
     "gulp-imagemin": "^7.1.0",
     "gulp-plumber": "1.2.x",
@@ -17,6 +18,7 @@
     "gulp-sass": "^4.0.2",
     "gulp-sourcemaps": "2.6.x",
     "gulp-svgstore": "^7.0.1",
+    "gulp-uglify": "^3.0.2",
     "gulp-webp": "^4.0.1",
     "node-sass": "^4.14.1",
     "posthtml-include": "^1.4.3",
@@ -47,6 +49,7 @@
     "node": "8.16"
   },
   "dependencies": {
-    "del": "^5.1.0"
+    "del": "^5.1.0",
+    "gulp-htmlmin": "^5.0.1"
   }
 }
