+++ source/catalog.html
@@ -257,8 +257,7 @@
     </section>
     <div class="overlay"></div>
 
-    <script src="js/main-nav.js"></script>
-    <script src="js/modal.js"></script>
+    <script src="js/main.min.js"></script>
     <script src="js/picturefill.min.js"></script>
     <script src="js/svg4everybody.min.js"></script>
