+++ source/form.html
@@ -269,7 +269,7 @@
       </div>
     </footer>
 
-    <script src="js/main-nav.js"></script>
+    <script src="js/main.min.js"></script>
     <script src="js/picturefill.min.js"></script>
     <script src="js/svg4everybody.min.js"></script>
