+++ source/js/main.js
@@ -1,3 +1,20 @@
+var navMain = document.querySelector('.main-nav');
+var navToggle = document.querySelector('.main-nav__toggle');
+
+navMain.classList.remove('main-nav--nojs');
+
+navToggle.addEventListener('click', function() {
+  if (navMain.classList.contains('main-nav--closed')) {
+    navMain.classList.remove('main-nav--closed');
+    navMain.classList.add('main-nav--opened');
+  } else {
+    navMain.classList.add('main-nav--closed');
+    navMain.classList.remove('main-nav--opened');
+  }
+});
+
+/*pop-up*/
+
 var cart = document.querySelectorAll(".js-button");
 var popup = document.querySelector(".modal-popup");
 var overlay = document.querySelector(".overlay");
